Consolidate z2_vison_higgs_confinement into z2_vison_higgs/
Split subcommand __main__.py into ground_state.py + dynamics.py using
runner. Preserves all Higgs physics: parity-sector model, interior vison
pair creation, all-plaquette observables for 2D map snapshots.
